Update to Scan function.
The Limit value was not used, so I added, ok. 
The param "ScanIndexForward" can be added like the Query function "$options" in the future. 
By default, the sort order is already ascending "true" by AWS. To reverse the order, set the ScanIndexForward parameter to false. 
Someday I will update using the same approach as $options = array() in query.